feat: auth

+++ controllers/cards.js
@@ -1,83 +1,52 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ReferenceError') {
-        return res.status(404).send({ message: err.message });
-      } return res.status(500).send({ message: 'Произошла ошибка' });
-    });
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(() => {
-      const err = new Error('Карточка не найдена');
-      err.name = 'not_found';
-      throw err;
+      throw new NotFoundError('Карточка не существует');
     })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданы неправильные данные' });
+    .then((card) => {
+      if (card.owner.toString() !== req.user._id) {
+        throw new ForbiddenError('Нельзя удалять чужие карточки');
       }
-      if (err.name === 'not_found') {
-        return res.status(404).send({ message: err.message });
-      }
-      return res.status(500).send({ message: 'Произошла ошибка' });
-    });
+      Card.findByIdAndRemove(req.params.cardId)
+        .then(() => res.send({ data: card }));
+    })
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const userId = req.user._id;
   Card.create({ name, link, owner: userId })
     .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: err.message });
-      }
-      return res.status(500).send({ message: 'Произошла ошибка' });
-    });
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+    { $addToSet: { likes: req.user._id } },
     { new: true },
   ).orFail(() => {
-    const err = new Error('id не найден');
-    err.name = 'not_found';
-    throw err;
-  }).then((card) => res.send({ data: card })).catch((err) => {
-    if (err.name === 'CastError') {
-      return res.status(400).send({ message: 'Переданы неправильные данные' });
-    }
-    if (err.name === 'not_found') {
-      return res.status(404).send({ message: err.message });
-    }
-    return res.status(500).send({ message: 'Произошла ошибка' });
-  });
+    throw new NotFoundError('_id не найден');
+  }).then((card) => res.send({ data: card })).catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $pull: { likes: req.user._id } }, // убрать _id из массива
+    { $pull: { likes: req.user._id } },
     { new: true },
   ).orFail(() => {
-    const err = new Error('id не найден');
-    err.name = 'not_found';
-    throw err;
-  }).then((card) => res.send({ data: card })).catch((err) => {
-    if (err.name === 'CastError') {
-      return res.status(400).send({ message: 'Переданы неправильные данные' });
-    }
-    if (err.name === 'not_found') {
-      return res.status(404).send({ message: err.message });
-    }
-    return res.status(500).send({ message: 'Произошла ошибка' });
-  });
+    throw new NotFoundError('_id не найден');
+  }).then((card) => res.send({ data: card })).catch(next);
 };
